--- src/components/CardInformation/CardInformation.tsx
@@ -1,30 +1,0 @@
-import CardLabel from '../CardLabel/CardLabel';
-import CardNumberInput from '../CardInput/CardInput';
-import styled from 'styled-components';
-
-interface CardInformationProps {
-  labelText: string;
-}
-
-const Wrapper = styled.div`
-  display: flex;
-  justify-content: center;
-  background: #ecebf1;
-  border-radius: 7px;
-`;
-
-const CardInformation = ({ labelText }: CardInformationProps) => {
-  return (
-    <>
-      <CardLabel labelText={labelText} />
-      <Wrapper>
-        <CardNumberInput type="text" maxLength={4} />
-        <CardNumberInput type="text" maxLength={4} />
-        <CardNumberInput type="text" maxLength={4} />
-        <CardNumberInput type="text" maxLength={4} />
-      </Wrapper>
-    </>
-  );
-};
-
-export default CardInformation;
